third_party/WebKit/LayoutTests/fast/css/background-currentcolor-expected.txt

Test that background-color is non-inherit and currentColor is handled correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(test1).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(test2).backgroundColor is "rgb(0, 128, 0)"
PASS getComputedStyle(test3).backgroundColor is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
